+++ src/components/custom-field-input-number/custom-field-input-number.jsx
@@ -1,46 +1,37 @@
 import PropTypes from 'prop-types';
 import React, { useEffect, useRef, useState } from 'react';
-
-import CustomFieldInputText from '../custom-field-input-text/custom-field-input-text.jsx';
-import styles from '../custom-field-input-text/custom-field-input-text.css';
+import styles from '../__internal__/abstract-custom-field.css';
+import useValidation from '../../hooks/use-validation.jsx';
+import AbstractCustomField from '../__internal__/abstract-custom-field.jsx';
 
 const apiLimits = {
   max: 2 ** 31,
   min: -(2 ** 31),
 };
 
-function getRootClassName(className, error, disabled) {
-  if (disabled) {
-    return `${className} ${styles.disabled}`;
-  }
-
-  if (error) {
-    return `${className} ${styles.error}`;
-  }
-
-  return className;
-}
-
 export default function CustomFieldInputNumber(props) {
-  const inputRef = props.inputRef || useRef(null);
-  const [invalid, setInvalid] = useState(false);
+  const fallbackRef = useRef(null);
+  const inputRef = props.inputRef || fallbackRef;
+
+  const [checkedValidity, setCheckedValidity] = useState(false);
+  const validationMessage = useValidation(props.readOnly, props.errorText, inputRef, checkedValidity);
 
   function handleOnKeyUp(event) {
-    setInvalid(!event.target.checkValidity());
+    setCheckedValidity(!event.target.checkValidity());
   }
 
   useEffect(() => {
     if (!inputRef.current) return;
 
-    setInvalid(!inputRef.current.validity.valid);
+    setCheckedValidity(!inputRef.current.validity.valid);
   });
 
   return (
-    <CustomFieldInputText
-      className={getRootClassName(props.className, invalid, props.disabled)}
+    <AbstractCustomField
+      className={props.className}
       defaultValue={props.value}
       disabled={props.disabled}
-      error={invalid}
+      errorText={validationMessage}
       id={props.id}
       inputRef={inputRef}
       label={props.label}
@@ -61,6 +52,7 @@ export default function CustomFieldInputNumber(props) {
 CustomFieldInputNumber.propTypes = {
   className: PropTypes.string,
   disabled: PropTypes.bool,
+  errorText: PropTypes.string,
   id: PropTypes.string.isRequired,
   inputRef: PropTypes.shape({ current: PropTypes.any }),
   label: PropTypes.string.isRequired,
@@ -80,6 +72,7 @@ CustomFieldInputNumber.propTypes = {
 CustomFieldInputNumber.defaultProps = {
   className: styles['custom-field-input-text'],
   disabled: false,
+  errorText: '',
   inputRef: undefined,
   name: undefined,
   onBlur: () => {},